src/WeatherForecastKPDL/Services/Backend/Prediction/WeatherForecast.Prediction/requirements.txt
apscheduler==3.10.1
celery==5.3.0
fastapi==0.109.1
httpx==0.26.0
joblib==1.3.2
kafka-python==2.0.2
numpy==1.26.3
pandas==2.2.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
redis[hiredis]==4.5.0
scikit-learn==1.4.0
uvicorn==0.27.0